changed int to long

+++ pagerank/parallel.c
@@ -148,7 +148,7 @@ void computeVector(long N, int p, int s, MPI_Comm comm)
         printf("Element %ld\n", outgoingLinks[i]);
     for(long r = 0; r < p; r++)
     {
-        MPI_Isend(outgoingLinks + outOffsets[r], outgoingDiagonal[r], MPI_INT, r, r, comm, &requests[r]);
+        MPI_Isend(outgoingLinks + outOffsets[r], outgoingDiagonal[r], MPI_LONG, r, r, comm, &requests[r]);
         MPI_Irecv(incoming + r * maxsize, sizes[r] , MPI_LONG, r, s, comm, &requests[p+r]);
     }
     MPI_Waitall(2*p, requests,MPI_STATUSES_IGNORE);
